Use rust `generate_new_device` function in python
